connect search to backend
