+++ components/search/SearchBarPage.jsx
@@ -1,34 +1,24 @@
 /* eslint-disable no-unused-vars */
 import React, { useState, useCallback, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import CourseData from '../../data/data';
 import TabBar from './TabBar';
 import SearchBar from './SearchBar';
 import SearchResults from './SearchResults';
 
 function SearchBarPage(props) {
+  const dispatch = useDispatch();
   const [searchInput, setSearchInput] = useState('');
-  const [tab, setTab] = useState('All');
+  const [tab, setTab] = useState('Courses');
   const [searchResults, setSearchResults] = useState(null);
   const { children } = props;
-
-  const input = useCallback((inputElement) => {
-    if (inputElement) {
-      inputElement.focus();
-    }
-  }, []);
-
-  const onBlur = (e) => {
-    window.setTimeout(() => e.target.focus(), 0);
-  };
 
   return (
     <div style={{ padding: '20px 80px 50px 275px' }}>
       <SearchBar
         searchInput={searchInput}
         setSearchInput={setSearchInput}
-        input={input}
         setSearchResults={setSearchResults}
-        onBlur={onBlur}
       />
       {
         searchInput // if search input exists, show tab bar page
