+++ styles/components/TabBar.module.css
@@ -32,6 +32,6 @@
     background: var(--pastel-pink);
 }
 
-.All {
+.Users {
     background: var(--pastel-blue);
 }
